Newlines in desc work. Added ruby sym file parser
Added a new css p class that preserves newlines.

Started a ruby program to parse sym files. At this point it can separate the equations and properties, and it makes guesses about what each property denotes based on the words used. The program has a list of key words for each field, then makes a guess about what each field is based on the key words.